add tests for factory, and modify for send notification data

+++ src/main/kotlin/com/untref/tesis/server/alert/domain/Coordinate.kt
@@ -1,3 +1,3 @@
 package com.untref.tesis.server.alert.domain
 
-data class Coordinate(val degree: Int, val minute: Int, val second: Float, val cardinalPoint: CardinalPoint)
+data class Coordinate(val degree: Int, val minute: Int, val second: Double, val cardinalPoint: CardinalPoint)
